+++ src/components/question/Question.js
@@ -1,4 +1,7 @@
 import React, {Component} from 'react';
+
+import DisplayQuestion from './DisplayQuestion'
+import DisplayAnswers from './DisplayAnswers'
 import VerbDisplayTable from '../verb/ui/VerbDisplayTable'
 
 //require data files
@@ -6,48 +9,9 @@ const rules = require('../../data/rules.json');
 
 class Question extends Component{
     
-    constructor(props){
-        super(props)
-        this.getDisplayQuestion = this.getDisplayQuestion.bind(this)
-        this.getAnswerList = this.getAnswerList.bind(this)
-    }
-
-    // returns display question str
-    getDisplayQuestion(pronoun, tense, verbName){
-        return (
-            <p className="questionText">Which is the correct <strong>{pronoun}</strong> form for the <strong>{tense}</strong> tense of <b>{verbName}</b>?</p>
-        )
-    }
-
-    getAnswerList(verbTablesArray, personIndex){
-        
-
-        const listItems = verbTablesArray.map((vt, idx) => {
-            let classNameString = 'questionText';    
-            classNameString += (this.getHasIsCorrectProp(vt)) ? ' correct': ''
-            
-            return <li key={idx} className={classNameString}>{vt[personIndex][1]}</li>
-        })
-        
-        return(
-            <ol>
-                {listItems}
-            </ol>
-        )
-
-    }
-
-    getHasIsCorrectProp(vt){
-        return vt[vt.length-1]['isCorrect']
-    }
-
-    
     render(){
-        const personIndex = this.props.params.personIndex
+        const {personIndex, tense, verbName, verbTablesArray} = this.props
         const pronoun = rules['pronouns'][personIndex]
-        const tense = this.props.params.tense
-        const verbName = this.props.params.verbName
-        const verbTablesArray = this.props.params.verbTablesArray
 
         if(verbName === "") {
             return(
@@ -55,23 +19,14 @@ class Question extends Component{
         )}
         return(           
             <div>
-                {this.getDisplayQuestion(pronoun, tense, verbName)}          
-                {this.getAnswerList(verbTablesArray, personIndex)}
-                {VerbDisplayTable(verbTablesArray[0])}
+                <DisplayQuestion pronoun={pronoun} tense={tense} verbName={verbName} />
+                <DisplayAnswers verbTablesArray={verbTablesArray} personIndex={personIndex} />
+                <VerbDisplayTable verbTablesArray={verbTablesArray} />
             </div>
         )
         
     }
 
-}
-
-Question.defaultProps = {
-    "params":{
-        "personIndex": 0,
-        "tense": "",
-        "vwerbName": ""
-    }
-    
 }
 
 export default Question
